+++ ChestsAnywhere/docs/release-notes.md
@@ -4,6 +4,7 @@
 ## Upcoming release
 * Fixed chest labels shown for chests at equivalent positions in another building of the same type.
 * Fixed 'current location' range including all buildings of the same type while inside one of them.
+* Fixed unable to paste into edit-chest textboxes.
 * Improved translations. Thanks to DiQuintino (updated Portuguese), JanUwU42 (updated German), and MakinDay (updated Italian)!
 
 ## 1.28.0